// src/types/whatsapp/webhook.ts
export type WebhookContact = {
  /**
   * The customer's WhatsApp ID. A business can respond to a message using this ID.
   */
  wa_id: string
  profile: {
    name: string
  }
}

export type WebhookError = {
  code: WABAErrorCodes
  title: string
}
// ...
export type WebhookMessage = {
// ...
  button?: {
    /**
     * The payload for a button set up by the business that a customer clicked as part of an interactive message
     */
    payload: any
    text: string
  }
// ...
  interactive?: {
    type: "button_reply" | "list_reply"
    button_reply?: {
      id: string
      title: string
    }
    list_reply?: {
      id: string
      title: string
      description: string
    }
  }
// ...
}

/**
 * For more information about this object, go here https://developers.facebook.com/docs/whatsapp/cloud-api/webhooks/components#statuses-object
 */
export type WebhookStatus = {
  /**
   * The ID for the message that the business that is subscribed to the webhooks sent to a customer
   */
  id: string
  /**
   * An object containing billing information.
   */
  pricing: {
    /**
     * Indicates the conversation pricing category
     */
    category: {
      /**
       * The business sent a message to a customer more than 24 hours after the last customer message
       */
      business_initiated: boolean
      /**
       * The conversation originated from a free entry point. These conversations are always customer-initiated.
       */
      referral_conversion: boolean
      /**
       * The business replied to a customer message within 24 hours of the last customer message
       */
      customer_initiated: boolean
    }
    /**
     *  Type of pricing model used by the business.
     */
    pricing_model: LiteralUnion<"CBP">
  }
  conversation: {
    /**
     *  Represents the ID of the conversation the given status notification belongs to.
     */
    id: string
    /**
     *  Indicates who initiated the conversation
     */
    origin: {
      /**
       * Indicates where a conversation has started. This can also be referred to as a conversation entry point
       */
      type: {
        /**
         * Indicates that the conversation started by a business sending the first message to a customer. This applies any time it has been more than 24 hours since the last customer message.
         */
        business_initiated: boolean
        /**
         * Indicates that the conversation started by a business replying to a customer message. This applies only when the business reply is within 24 hours of the last customer message.
         */
        customer_initiated: boolean
        /**
         * Indicates that the conversation originated from a free entry point. These conversations are always customer-initiated.
         */
        referral_conversation: boolean
      }
      /**
       *  Date when the conversation expires. This field is only present for messages with a `status` set to `sent`
       */
      expiration_timestamp: string
    }
  }
  /**
   * The WhatsApp ID for the customer that the business, that is subscribed to the webhooks, sent to the customer
   */
  recipient_id: string
  status: "delivered" | "failed" | "read" | "sent"
  /**
   * Date for the status message in unix
   */
  timestamp: number
}

export type WebhookMetadata = {
  display_phone_number: string
  phone_number_id: string
}

export type WebhookChange = {
  value: {
    messaging_product: "whatsapp"
    metadata: WebhookMetadata
    errors: WebhookError[]
    contacts: WebhookContact[]
    messages?: WebhookMessage[]
    statuses?: WebhookStatus[]
  }
  field: LiteralUnion<"messages">
}

/**
 * Webhooks are triggered when a customer performs an action or the status for a message a business sends a customer changes.
 * To add webhooks go here https://developers.facebook.com/docs/whatsapp/cloud-api/guides/set-up-webhooks
 * To see examples of Webhooks Responses go here: https://developers.facebook.com/docs/whatsapp/cloud-api/webhooks/payload-examples
 */
export type Webhook = {
  object: string
  entry: [
    {
      id: string
      changes: WebhookChange[]
    }
  ]
}